#94 renamed multiple sources to several targets - which is a more suitable name -
